requirements.txt
allure-pytest==2.13.5
allure-python-commons==2.13.5
attrs==23.2.0
certifi==2024.2.2
click==8.1.7
decorator==5.1.1
docutils==0.21.2
driver==0.1.0
h11==0.14.0
idna==3.7
iniconfig==2.0.0
lockfile==0.12.2
outcome==1.3.0.post0
packaging==24.0
pid==3.0.4
pluggy==1.5.0
PySocks==1.7.1
pytest==8.2.0
python-daemon==3.0.1
selenium==4.20.0
self==2020.12.3
service==0.6.0
setproctitle==1.3.3
setuptools==69.5.1
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.25.0
trio-websocket==0.11.1
typing_extensions==4.11.0
urllib3==2.2.1
wsproto==1.2.0
